feat(sdk): add service-worker dedicated target
includes only functions to create client from stringified type
and download files and zips to a stream.
This shrinks the binary size for including the wasm in a service worker
to ~850KiB from ~5.4MiB.
Additionally with this approach
the service worker doesn't run into issues with multi-threading,
as spawning web workers is disallowed in as service worker.